Upgrade Laravel core from 7.0 to 8.0
