+++ database/seeders/AdministradorSeeder.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\User;
 use Carbon\Carbon;
 use Illuminate\Database\Seeder;
