+++ database/seeders/ProfessoresSeeder.php
@@ -1,37 +1,37 @@
 <?php
 
-use App\Aluno;
+namespace Database\Seeders;
+
 use App\Nucleo;
+use App\Professores;
 use App\User;
 use Carbon\Carbon;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\Hash;
 
-class AlunosSeeder extends Seeder
+class ProfessoresSeeder extends Seeder
 {
 
     public function run() {
-        $role = 'aluno';
+        $role = 'professor';
 
         $data = [
-            'Aluno A' => [
-                'name' => 'Aluno A',
-                'email' => 'alunoa@biko.edu',
-                'password' => 'alunoa@biko.edu',
+            'Professor A' => [
+                'name' => 'Professor A',
+                'email' => 'Professora@biko.edu',
+                'password' => 'Professora@biko.edu',
                 'phone' => '6430000000',
                 'nucleo' => 'Núcleo A',
                 'status' => 1,
-                'listaEspera' => 'Não',
                 'emailVerified' => Carbon::now(),
             ],
-            'Aluno B' => [
-                'name' => 'Aluno B',
-                'email' => 'alunob@biko.edu',
-                'password' => 'alunob@biko.edu',
+            'Professor B' => [
+                'name' => 'Professor B',
+                'email' => 'Professorb@biko.edu',
+                'password' => 'Professorb@biko.edu',
                 'phone' => '6440000000',
                 'nucleo' => 'Núcleo B',
                 'status' => 0,
-                'listaEspera' => 'Sim',
                 'emailVerified' => null,
             ],
         ];
@@ -48,15 +48,13 @@ class AlunosSeeder extends Seeder
 
             $nucleo = Nucleo::where('NomeNucleo', $row['nucleo'])->first();
 
-            Aluno::create([
-                'NomeAluno' => $user->name,
+            Professores::create([
+                'NomeProfessor' => $user->name,
                 'id_user' => $user->id,
                 'Status' => $row['status'],
                 'FoneCelular' => $user->phone,
                 'Email' => $user->email,
                 'id_nucleo' => $nucleo->id,
-                'NomeNucleo' => $nucleo->NomeNucleo,
-                'ListaEspera' => $row['listaEspera'],
             ]);
         }
     }
